+++ app/src/ingest_bem_pdfs.py
@@ -9,7 +9,7 @@ from smart_open import open as smart_open
 from src.adapters import db
 from src.app_config import app_config
 from src.db.models.document import Chunk, Document
-from src.ingestion.get_grouped_texts import get_grouped_texts
+from src.ingestion.pdf_postprocess import group_texts
 from src.ingestion.pdf_elements import EnrichedText
 from src.util import pdf_utils
 from src.util.file_util import get_files
